add container atom and make all this stateless
