+++ ui/index.js
@@ -1,2 +1,3 @@
 export * from './components/atoms/Button'
+export * from './components/atoms/Container'
 export * from './components/atoms/Logo'
